Refactor store.py into two separate files
